requirements.txt
pybullet
gymnasium
stable-baselines3
numpy
matplotlib
pillow